+++ syndicate/core/handlers.py
@@ -129,10 +129,16 @@ def syndicate():
 @click.option('--errors_allowed', is_flag=True,
               help='Flag to return successful response even if some tests '
                    'fail')
+@click.option('--skip_tests', is_flag=True, default=False,
+              help='Flag to not run tests')
 @verbose_option
 @timeit(action_name=TEST_ACTION)
-def test(suite, test_folder_name, errors_allowed):
+def test(suite, test_folder_name, errors_allowed, skip_tests):
     """Discovers and runs tests inside python project configuration path."""
+    if skip_tests:
+        click.echo('Skipping tests...')
+        return
+
     click.echo('Running tests...')
     import subprocess
     from syndicate.core import CONFIG
@@ -160,11 +166,14 @@ def test(suite, test_folder_name, errors_allowed):
     if result.returncode != 0:
         _LOG.error(f'{result.stdout}\n{result.stderr}\n{"-" * 70}')
         if not errors_allowed:
-            click.echo('Some tests failed. Exiting.')
+            click.echo(
+                'Some tests failed. See details in the log file. Exiting...')
             sys.exit(result.returncode)
-
-    _LOG.info(f'{result.stdout}\n{result.stderr}\n{"-" * 70}')
-    click.echo('Tests passed.')
+        else:
+            USER_LOG.warn('Some tests failed. See details in the log file.')
+    else:
+        _LOG.info(f'{result.stdout}\n{result.stderr}\n{"-" * 70}')
+        click.echo('Tests passed.')
 
 
 @syndicate.command(name=BUILD_ACTION)
@@ -175,12 +184,15 @@ def test(suite, test_folder_name, errors_allowed):
 @click.option('--force_upload', '-F', is_flag=True, default=False,
               help='Flag to override existing bundle with the same name')
 @click.option('--errors_allowed', is_flag=True, default=False,
-              help='Flag to continue bundle building if some tests fail')
+              help='Flag to continue building the bundle if any errors occur '
+                   'while building dependencies or tests fail')
+@click.option('--skip_tests', is_flag=True, default=False,
+              help='Flag to skip lambda tests')
 @verbose_option
 @click.pass_context
 @check_deploy_bucket_exists
 @timeit(action_name=BUILD_ACTION)
-def build(ctx, bundle_name, force_upload, errors_allowed):
+def build(ctx, bundle_name, force_upload, errors_allowed, skip_tests):
     """
     Builds bundle of an application
     """
@@ -189,8 +201,9 @@ def build(ctx, bundle_name, force_upload, errors_allowed):
                    f'in deploy bucket. Please use another bundle '
                    f'name or delete the bundle')
         return
-    ctx.invoke(test, errors_allowed=errors_allowed)
-    ctx.invoke(assemble, bundle_name=bundle_name, force_upload=force_upload)
+    ctx.invoke(test, errors_allowed=errors_allowed, skip_tests=skip_tests)
+    ctx.invoke(assemble, bundle_name=bundle_name,
+               errors_allowed=errors_allowed, force_upload=force_upload)
     ctx.invoke(package_meta, bundle_name=bundle_name)
     ctx.invoke(upload, bundle_name=bundle_name, force_upload=force_upload)
 
@@ -638,9 +651,12 @@ def profiler(bundle_name, deploy_name, from_date, to_date):
               help='Identifier that indicates whether a locally existing'
                    ' bundle should be deleted and a new one created using'
                    ' the same path.')
+@click.option('--skip_tests', is_flag=True, default=False,
+              help='Flag to not run tests')
 @verbose_option
 @timeit(action_name=ASSEMBLE_JAVA_MVN_ACTION)
-def assemble_java_mvn(bundle_name, project_path, force_upload):
+def assemble_java_mvn(bundle_name, project_path, force_upload, skip_tests,
+                      errors_allowed=False):
     """
     Builds Java lambdas
 
@@ -648,14 +664,17 @@ def assemble_java_mvn(bundle_name, project_path, force_upload):
     :param bundle_name: name of the bundle
     :param project_path: path to project folder
     :param force_upload: force upload identification
+    :param skip_tests: force skipping tests
+    :param errors_allowed: not used for java, but need to unify the
+    `assemble` commands interface
     :return:
     """
     click.echo(f'Command compile java project path: {project_path}')
     assemble_artifacts(bundle_name=bundle_name,
                        project_path=project_path,
                        runtime=RUNTIME_JAVA,
-                       force_upload=force_upload
-                       )
+                       force_upload=force_upload,
+                       skip_tests=skip_tests)
     click.echo('Java artifacts were prepared successfully.')
 
 
@@ -678,9 +697,13 @@ def assemble_java_mvn(bundle_name, project_path, force_upload):
               help='Identifier that indicates whether a locally existing'
                    ' bundle should be deleted and a new one created using'
                    ' the same path.')
+@click.option('--errors_allowed', is_flag=True, default=False,
+              help='Flag to continue building the bundle if any errors occur '
+                   'while building dependencies')
 @verbose_option
 @timeit(action_name=ASSEMBLE_PYTHON_ACTION)
-def assemble_python(bundle_name, project_path, force_upload):
+def assemble_python(bundle_name, project_path, force_upload, errors_allowed,
+                    skip_tests=False):
     """
     Builds Python lambdas
 
@@ -688,14 +711,17 @@ def assemble_python(bundle_name, project_path, force_upload):
     :param bundle_name: name of the bundle
     :param project_path: path to project folder
     :param force_upload: force upload identification
+    :param errors_allowed: allows to ignore dependency errors
+    :param skip_tests: not used for python, but need to unify the
+    `assemble` commands interface
     :return:
     """
     click.echo(f'Command assemble python: project_path: {project_path} ')
     assemble_artifacts(bundle_name=bundle_name,
                        project_path=project_path,
                        runtime=RUNTIME_PYTHON,
-                       force_upload=force_upload
-                       )
+                       force_upload=force_upload,
+                       errors_allowed=errors_allowed)
     click.echo('Python artifacts were prepared successfully.')
 
 
@@ -718,7 +744,8 @@ def assemble_python(bundle_name, project_path, force_upload):
                    ' the same path.')
 @verbose_option
 @timeit(action_name=ASSEMBLE_NODE_ACTION)
-def assemble_node(bundle_name, project_path, force_upload):
+def assemble_node(bundle_name, project_path, force_upload,
+                  errors_allowed=False, skip_tests=False):
     """
     Builds NodeJS lambdas
 
@@ -726,6 +753,10 @@ def assemble_node(bundle_name, project_path, force_upload):
     :param bundle_name: name of the bundle
     :param project_path: path to project folder
     :param force_upload: force upload identification
+    :param errors_allowed: not used for NodeJS, but need to unify the
+    `assemble` commands interface
+    :param skip_tests: not used for NodeJS, but need to unify the
+    `assemble` commands interface
     :return:
     """
     click.echo(f'Command assemble node: project_path: {project_path} ')
@@ -756,7 +787,8 @@ def assemble_node(bundle_name, project_path, force_upload):
                    ' the same path.')
 @verbose_option
 @timeit(action_name=ASSEMBLE_DOTNET_ACTION)
-def assemble_dotnet(bundle_name, project_path, force_upload):
+def assemble_dotnet(bundle_name, project_path, force_upload,
+                    errors_allowed=False, skip_tests=False):
     """
     Builds DotNet lambdas
 
@@ -764,6 +796,10 @@ def assemble_dotnet(bundle_name, project_path, force_upload):
     :param bundle_name: name of the bundle
     :param project_path: path to project folder
     :param force_upload: force upload identification
+    :param errors_allowed: not used for DotNet, but need to unify the
+    `assemble` commands interface
+    :param skip_tests: not used for DotNet, but need to unify the
+    `assemble` commands interface
     :return:
     """
     click.echo(f'Command assemble dotnet: project_path: {project_path} ')
@@ -824,10 +860,13 @@ RUNTIME_LANG_TO_BUILD_MAPPING = {
               help='Identifier that indicates whether a locally existing'
                    ' bundle should be deleted and a new one created using'
                    ' the same path.')
+@click.option('--errors_allowed', is_flag=True, default=False,
+              help='Flag to continue building the bundle if any errors occur '
+                   'while building dependencies. Only for Python runtime.')
 @verbose_option
 @click.pass_context
 @timeit(action_name=ASSEMBLE_ACTION)
-def assemble(ctx, bundle_name, force_upload):
+def assemble(ctx, bundle_name, force_upload, errors_allowed):
     """
     Builds the application artifacts
 
@@ -835,6 +874,8 @@ def assemble(ctx, bundle_name, force_upload):
     :param ctx:
     :param bundle_name: name of the bundle to which the artifacts
         will be associated
+    :param force_upload: force upload identification
+    :param errors_allowed: allows to ignore dependency errors. Only for python
     :return:
     """
     click.echo(f'Building artifacts, bundle: {bundle_name}')
@@ -848,7 +889,8 @@ def assemble(ctx, bundle_name, force_upload):
             func = RUNTIME_LANG_TO_BUILD_MAPPING.get(key)
             if func:
                 ctx.invoke(func, bundle_name=bundle_name,
-                           project_path=value, force_upload=force_upload)
+                           project_path=value, force_upload=force_upload,
+                           errors_allowed=errors_allowed)
             else:
                 click.echo(f'Build tool is not supported: {key}')
     else: