+++ CHANGELOG.md
@@ -4,7 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-# [1.14.2] - 2024-10-25
+# [1.15.0] - 2024-10-28
+- Added `--skip_tests` option to `build`, `test` and `assemble_java_mvn` commands to not run tests during or after 
+building the bundle
+- Added `--errors_allowed` option to `assemble_python` and `assemble` commands
+- Improved logic of option `--errors_allowed` to allow dependency installing if it is not possible to find a 
+requirement that fits a specific platform. In this case, the dependencies will be installed independently 
+of each other or without specifying a specific platform (using default platform `any`). Python-specific feature
 - Fixed updating lambda layers when the lambda no longer has layers
 - Fixed logging not found exceptions during the clean operation
 - Fixed handling of deployment output in case of failures on the stage describe resources in case of deploy/update fail